+++ src/FsUnit.Xunit/FsUnit.fs
@@ -26,6 +26,17 @@ let inline should (f : 'a -> ^b) x (y : obj) =
         | _ -> y
     Assert.That(y, c)
 
+let inline shouldFail (f:unit->unit) =
+    let failed =
+        try
+            f()
+            false
+        with
+        | _ -> true
+    if not failed then
+        raise (new MatchException("Method should fail", "No exception raised", null))
+
+
 let equal expected = CustomMatchers.equal expected
 
 let equalWithin (tolerance:obj) (expected:obj) = CustomMatchers.equalWithin tolerance expected