+++ tests/FsUnit.MsTest.Test/shouldFailTests.fs
@@ -1,27 +1,28 @@
 ﻿namespace FsUnit.Test
-open NUnit.Framework
-open FsUnit
+open Microsoft.VisualStudio.TestTools.UnitTesting
+open FsUnit.MsTest
+open NHamcrest.Core
 
-[<TestFixture>]
+[<TestClass>]
 type ``shouldFail tests`` ()=
-    [<Test>] member test.
+    [<TestMethod>] member test.
      ``empty List should fail to contain item`` ()=
         shouldFail (fun () -> [] |> should contain 1)
 
-    [<Test>] member test.
+    [<TestMethod>] member test.
      ``non-null should fail to be  Null`` ()=
         shouldFail (fun () -> "something" |> should be Null)
 
-    [<Test>] member test.
+    [<TestMethod>] member test.
      ``shouldFail should fail when everything is OK`` ()=
         shouldFail (fun () -> shouldFail id)
 
-    [<Test>] member test.
+    [<TestMethod>] member test.
      ``shouldFaild should throw an exception`` ()=
         (fun () -> shouldFail id)
-        |> should throw typeof<NUnit.Framework.AssertionException>
+        |> should throw typeof<AssertFailedException>
 
-    [<Test>] member test.
+    [<TestMethod>] member test.
      ``shouldFaild should not throw an exception when fail`` ()=
         (fun () -> shouldFail (fun () -> [] |> should contain 1))
-        |> should not' (throw typeof<NUnit.Framework.AssertionException>)
+        |> should not' (throw typeof<AssertFailedException>)