+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 * FsUnitDepricated module renamed to FsUnitDeprecated https://github.com/fsprojects/FsUnit/pull/46
 * Enable substring checks https://github.com/fsprojects/FsUnit/pull/45
 * Fixed assertion message for NUnit equal constraint - https://github.com/fsprojects/FsUnit/pull/60
+* Added support of `shouldFail` operator for xUnit, MsTest and MbUnit
 
 ### 1.3.1.0 - 26 July 2015
 * Bump NUnit version up to 2.6.4